# JSON.generate: call to_json on String subclasses