Fix: https://github.com/ruby/json/issues/667

This is yet another behavior on which the various implementations differed, but the C implementation used to call `to_json` on String subclasses used as keys.

This was optimized out in e125072130229e54a651f7b11d7d5a782ae7fb65 but there is an Active Support test case for it, so it's best to make all 3 implementation respect this behavior.

FYI: @mtasaka