@eregon 2024-10-31
This seems a bit unfortunate performance-wise.
I wonder why `EscapedString` inherits from String, if it didn't it would just work without special care, no?