@byroot 2024-10-31
I think the performance cost is very minor, just an extra pointer comparison on the class, that's tagged as unlikely, so probably predicted out.

I didn't see much change on the benchmarks.