@eregon 2024-10-31
In lib/json/pure/generator.rb `generate_json` it seems quite a bit of extra code and checks, which seems likely to hurt dump perf quite a bit.