@eregon 2024-10-31
Before this PR (6d3b3ac):
```
$ ruby benchmark/standalone.rb dump pure
JSON::Pure::Generator
truffleruby 24.1.1, like ruby 3.2.4, Oracle GraalVM JVM [x86_64-linux]
Warming up --------------------------------------
      JSON.dump(obj)   350.000 i/100ms
      JSON.dump(obj)   745.000 i/100ms
      JSON.dump(obj)   829.000 i/100ms
      JSON.dump(obj)   799.000 i/100ms
      JSON.dump(obj)   836.000 i/100ms
Calculating -------------------------------------
      JSON.dump(obj)      8.152k (± 9.8%) i/s  (122.66 μs/i) -     40.128k in   5.008035s
      JSON.dump(obj)      8.166k (± 9.3%) i/s  (122.46 μs/i) -     40.964k in   5.084684s
      JSON.dump(obj)      8.180k (± 6.6%) i/s  (122.26 μs/i) -     40.964k in   5.038216s
      JSON.dump(obj)      8.101k (± 9.2%) i/s  (123.43 μs/i) -     40.128k in   5.020143s
      JSON.dump(obj)      8.156k (± 7.0%) i/s  (122.60 μs/i) -     40.964k in   5.056036s

$ ruby benchmark/standalone.rb dump pure
JSON::Pure::Generator
truffleruby 24.1.1, like ruby 3.2.4, Oracle GraalVM JVM [x86_64-linux]
Warming up --------------------------------------
      JSON.dump(obj)   336.000 i/100ms
      JSON.dump(obj)   793.000 i/100ms
      JSON.dump(obj)   827.000 i/100ms
      JSON.dump(obj)   828.000 i/100ms
      JSON.dump(obj)   788.000 i/100ms
Calculating -------------------------------------
      JSON.dump(obj)      8.074k (± 8.6%) i/s  (123.85 μs/i) -     40.188k in   5.037323s
      JSON.dump(obj)      7.993k (± 9.8%) i/s  (125.11 μs/i) -     39.400k in   5.004141s
      JSON.dump(obj)      8.032k (± 9.8%) i/s  (124.51 μs/i) -     40.188k in   5.079140s
      JSON.dump(obj)      8.013k (± 8.8%) i/s  (124.79 μs/i) -     40.188k in   5.073014s
      JSON.dump(obj)      8.098k (± 9.2%) i/s  (123.49 μs/i) -     40.188k in   5.038510s
```
After this PR (96397cf0903d4321263243feb8fb182e4698b6d9):
```
$ ruby benchmark/standalone.rb dump pure
JSON::Pure::Generator
truffleruby 24.1.1, like ruby 3.2.4, Oracle GraalVM JVM [x86_64-linux]
Warming up --------------------------------------
      JSON.dump(obj)   363.000 i/100ms
      JSON.dump(obj)   557.000 i/100ms
      JSON.dump(obj)   647.000 i/100ms
      JSON.dump(obj)   605.000 i/100ms
      JSON.dump(obj)   580.000 i/100ms
Calculating -------------------------------------
      JSON.dump(obj)      6.405k (± 7.1%) i/s  (156.14 μs/i) -     31.900k in   5.016689s
      JSON.dump(obj)      6.387k (± 9.6%) i/s  (156.57 μs/i) -     31.900k in   5.077313s
      JSON.dump(obj)      6.437k (± 5.6%) i/s  (155.34 μs/i) -     32.480k in   5.067145s
      JSON.dump(obj)      6.394k (± 8.5%) i/s  (156.40 μs/i) -     31.900k in   5.045133s
      JSON.dump(obj)      6.402k (± 7.7%) i/s  (156.20 μs/i) -     31.900k in   5.022605s

$ ruby benchmark/standalone.rb dump pure
JSON::Pure::Generator
truffleruby 24.1.1, like ruby 3.2.4, Oracle GraalVM JVM [x86_64-linux]
Warming up --------------------------------------
      JSON.dump(obj)   311.000 i/100ms
      JSON.dump(obj)   557.000 i/100ms
      JSON.dump(obj)   617.000 i/100ms
      JSON.dump(obj)   579.000 i/100ms
      JSON.dump(obj)   589.000 i/100ms
Calculating -------------------------------------
      JSON.dump(obj)      6.444k (± 7.8%) i/s  (155.18 μs/i) -     32.395k in   5.069218s
      JSON.dump(obj)      6.436k (± 7.9%) i/s  (155.37 μs/i) -     32.395k in   5.073131s
      JSON.dump(obj)      6.521k (± 6.3%) i/s  (153.36 μs/i) -     32.984k in   5.083403s
      JSON.dump(obj)      6.534k (± 5.7%) i/s  (153.04 μs/i) -     32.984k in   5.071127s
      JSON.dump(obj)      6.483k (±10.8%) i/s  (154.24 μs/i) -     31.806k in   5.023239s
```