@eregon 2024-10-31
I tried to undo the changes incrementally and the main slowdown seems from using:
```ruby
klass = obj.class
if klass == Hash
elsif klass == Array
elsif klass == String
...
```
(I also tried to reverse that == check but it didn't seem to change perf much, the reasoning is `klass->klass` is polymorphic, it's different singleton classes)

vs
```ruby
case obj
when Hash
when Array
when String
...
```