@byroot 2024-10-31
And how fast is the C implementation on Truffle?

Because on my machine:

```
$ ruby --yjit -Ilib:ext benchmark/standalone.rb dump
JSON::Ext::Generator
ruby 3.3.4 (2024-07-09 revision be1089c8ec) +YJIT [arm64-darwin23]
Warming up --------------------------------------
      JSON.dump(obj)     3.295k i/100ms
Calculating -------------------------------------
      JSON.dump(obj)     33.306k (± 0.6%) i/s   (30.02 μs/i) -    168.045k in   5.045675s
```

So maybe Truffle should just revert back to use the C extension?